requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

python-dotenv==1.0.1

SQLAlchemy==2.0.25
asyncpg==0.29.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

langchain==0.2.1
langchain-openai==0.1.7
langchain-core==0.2.3
openai>=1.24.0,<2.0.0

ragas==0.1.7
datasets==2.18.0

loguru==0.7.2
httpx==0.26.0
pydantic==2.6.1
